+++ src/main/java/ru/vyarus/guice/persist/orient/finder/result/ResultDesc.java
@@ -1,5 +1,7 @@
 package ru.vyarus.guice.persist.orient.finder.result;
 
+import ru.vyarus.guice.persist.orient.finder.internal.result.ResultDescriptor;
+
 /**
  * Result description for conversion.
  *
@@ -15,4 +17,14 @@ public class ResultDesc {
     // collection generic or method return if single result
     public Class entityClass;
 
+    public ResultDesc() {
+        // for tests
+    }
+
+    public ResultDesc(final ResultDescriptor descriptor, final Object result) {
+        type = descriptor.returnType;
+        returnClass = descriptor.expectType;
+        entityClass = descriptor.entityType;
+        this.result = result;
+    }
 }