+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/DocumentDao.groovy
@@ -3,12 +3,16 @@ package ru.vyarus.guice.persist.orient.support.finder
 import com.google.inject.persist.Transactional
 import com.orientechnologies.orient.core.record.impl.ODocument
 import ru.vyarus.guice.persist.orient.support.finder.mixin.crud.DocumentCrudMixin
+import ru.vyarus.guice.persist.orient.support.finder.mixin.pagination.PaginationMixin
+import ru.vyarus.guice.persist.orient.support.model.Model
 
 /**
  * @author Vyacheslav Rusakov 
  * @since 26.10.2014
  */
 @Transactional
-public interface DocumentDao extends DocumentCrudMixin, CustomMixin<ODocument, String> {
+public interface DocumentDao extends DocumentCrudMixin,
+        CustomMixin<ODocument, String>,
+        PaginationMixin<Model, ODocument> {
 
 }