+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/ObjectDao.groovy
@@ -2,6 +2,7 @@ package ru.vyarus.guice.persist.orient.support.finder
 
 import com.google.inject.persist.Transactional
 import ru.vyarus.guice.persist.orient.support.finder.mixin.crud.ObjectCrudMixin
+import ru.vyarus.guice.persist.orient.support.finder.mixin.pagination.PaginationMixin
 import ru.vyarus.guice.persist.orient.support.model.Model
 
 /**
@@ -9,5 +10,7 @@ import ru.vyarus.guice.persist.orient.support.model.Model
  * @since 26.10.2014
  */
 @Transactional
-interface ObjectDao extends ObjectCrudMixin<Model>, CustomMixin<Model, String> {
+interface ObjectDao extends ObjectCrudMixin<Model>,
+        CustomMixin<Model, String>,
+        PaginationMixin<Model, Model> {
 }