+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * AutoScanFinderModule accept multiple packages for scanning
 * Add experimental(!) delegate finders (annotated interface method delegates execution to guice bean method)
 * Add crud mixins for object and document finders (DocumentCrudMixin, ObjectCrudMixin)
+* Add document and vertex projection for single field (works for plain and array result): useful for count (or other aggregation functions) or selecting single field
+* Add simple pagination mixin (for object and document finders): PaginationMixin   
 
 ### 1.1.1 (2014-10-01)
 * Fix finder behaviour: empty collection conversion to single element